// generated-api/models/PredictionResponse.ts
export interface PredictionResponse {
    /**
     * 
     * @type {number}
     * @memberof PredictionResponse
     */
    predictionId?: number;
    /**
     * 
     * @type {string}
     * @memberof PredictionResponse
     */
    boxerAName?: string;
    /**
     * 
     * @type {string}
     * @memberof PredictionResponse
     */
    boxerBName?: string;
    /**
     * 
     * @type {string}
     * @memberof PredictionResponse
     */
    predictedWinner?: string;
    /**
     * 
     * @type {number}
     * @memberof PredictionResponse
     */
    weightClassId?: number;
    /**
     * 
     * @type {number}
     * @memberof PredictionResponse
     */
    closenessA?: number;
    /**
     * 
     * @type {number}
     * @memberof PredictionResponse
     */
    closenessB?: number;
    /**
     * 
     * @type {number}
     * @memberof PredictionResponse
     */
    probabilityA?: number;
    /**
     * 
     * @type {number}
     * @memberof PredictionResponse
     */
    probabilityB?: number;
    /**
     * 
     * @type {string}
     * @memberof PredictionResponse
     */
    explanation?: string;
    /**
     * 
     * @type {Date}
     * @memberof PredictionResponse
     */
    predictionDate?: Date;
}